requirements.txt
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
PyGithub>=2.1.0
anthropic>=0.40.0
markdown>=3.5.0
